Chapter 5 halfway complete.
Migrations for adding 'user' table to the database and adding 'auth_key' column to it.
ActiveRecord model and CRUD generated for 'user' table.
IdentityClass added to web.php config file.
dump.sql generated for functional tests.
Functional tests have not been implemented due to there being too many structural differences between what is in the book and what is in the current version of Yii2.